+++ src/shared/ringbuffer.cpp
