+++ src/shared/ringbuffer.h
