+++ src/usb-mic/audiodev-wasapi.h
@@ -2,7 +2,7 @@
 
 #include "audiodeviceproxy.h"
 #include "../libsamplerate/samplerate.h"
-#include "ringbuffer.h"
+#include "shared/ringbuffer.h"
 
 #include <mmdeviceapi.h>
 #include <audioclient.h>
